fix instace create

+++ app/Console/Commands/Whatsapp/SyncWhatsappInstances.php
@@ -46,11 +46,11 @@ class SyncWhatsappInstances extends Command
                     'token' => $instance->token,
                     'status' => InstanceStatus::NOT_AUTHORIZED,
                 ]);
-            } elseif (! $exists->status->isAuthorized()) {
+            } elseif (! $exists->status->isAuthorized() && $exists->created_at->lessThan(now()->subHours(12))) {
                 $exists->phone = null;
                 $exists->settings()->delete();
 
-                if ($exists->created_at->lessThan(now()->subHours(12)) && $exists->user_id) {
+                if ($exists->user_id) {
                     do_log('crones/instances')->warning("Инстанс {$exists->id} откреплень от клиента {$exists->user_id}");
                     $exists->user_id = null;
                 }
